Type fix (#726)
* fix: remove un-used types

* fix: change js to ts file
