+++ vite.config.ts
@@ -14,7 +14,7 @@ const config = {
 	},
 	resolve: {
 		alias: {
-			'flowbite-svelte': path.resolve(process.cwd(), './src/lib/index.js')
+			'flowbite-svelte': path.resolve(process.cwd(), './src/lib/index.ts')
 		}
 	}
 };
